Week 5 Coding Assignment -
menu driven application backed by a MySQL database. Error with updating life status and/or name information. Can enter data but does not update records and throws and error message depending on how the sql query statement is written. Narrowed the update option to just the life status update so only working with one variable instead of four with the name. Receive the same error with either update options.  Rest of the app works great!